requirements.txt
django==4.1.6
django-debug-toolbar==3.8.1
Pillow==9.4.0
environs[django]==9.5.0
django-phonenumber-field[phonenumberslite]==7.0.2
django-phonenumbers==1.0.1
djangorestframework==3.14.0
requests==2.28
geopy==2.3.0
